pyproject.toml
[project]
name = "MadokaBot"
version = "0.1.0"
description = "MadokaBot"
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "nonebot2[aiohttp,fastapi,websockets]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-console>=0.9.0",
    "nonebot-plugin-status>=0.9.0",
    "nonebot-plugin-alconna>=0.60.3",
    "nonebot-plugin-htmlrender>=0.6.7",
    "nonebot-plugin-datastore>=1.3.1",
    "sqlalchemy>=2.0.45",
    "Pillow>=10.2.0",
    "pytz>=2024.2",
    "numpy>=1.24.4",
]

[tool.nonebot]
plugin_dirs = []
builtin_plugins = []
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapter.onebot.v11" },
    { name = "Console", module_name = "nonebot.adapter.console" }
]
plugins = [
    "nonebot_plugin_htmlrender",
    "nonebot_plugin_apscheduler",
    "nonebot_plugin_status"
]
